app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
qrcode[pil]==7.4.2
pyyaml==6.0.2
python-multipart==0.0.10
